+++ src/library/Pqi/Conformance/Operation/SendQuery.hs
@@ -9,11 +9,14 @@ where
 import qualified Pqi
 import Pqi.Conformance.Harness
 import Pqi.Conformance.Prelude
+import qualified Pqi.Conformance.Operation.SendQuery.ConnectionLostBeforeSend as ConnectionLostBeforeSend
 import Pqi.Conformance.Scenario (drainResults)
 import Test.Hspec
 
 spec :: Pqi.Adapter -> SpecWith ByteString
-spec adapter =
+spec adapter = do
+  ConnectionLostBeforeSend.spec adapter
+
   describe "sendQuery" do
     let sendAndDrain sql conninfo =
           differential adapter conninfo \connection -> do